Move proxy_html.py to proxy.py
